Make the thresholds for the tap detection configurable
The current thresholds can easily cause conflicts with other gestures.

A maximum distance of 0.4f means you can swipe on the trackpad and
still trigger a middle click (the position is normalized and it is
always in the range [0..1]). Lower values, such as 0.05f, work better
for detecting just taps.

A maximum duration of 400ms is also quite a lot for a tap. Something
between 150ms/200ms works better for detecting taps.

Instead of changing the default values, which maybe work for some, this
adds two new preferences that allow to change those thresholds.

For example, to set the maximum distance to 0.05f and the maximum duration
to 200, run:

 defaults write com.rouge41.middleClick maxDistanceDelta 0.05
 defaults write com.rouge41.middleClick maxTimeDelta 200